+++ app/Http/Controllers/ServicesController.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Google_Service_Classroom;
+use Illuminate\Http\Request;
+use PulkitJalan\Google\Facades\Google;
+
+class ServicesController extends Controller
+{
+    public function drive() {
+        // $classroom = Google::make('classroom');
+
+        // $client = new \PulkitJalan\Google\Client($config);
+        // $googleClient = $client->getClient();
+
+    }
+}