+++ resources/lang/ca/challenges.php
@@ -43,4 +43,6 @@ return [
     'mark_confirm' => 'Confirme superació',
     'mark_text' => 'Estàs segur/1 que has superat la tasca? La confiança és necessària.',
 
+    'comment' => 'Comenta',
+
 ];