+++ resources/lang/en/challenges.php
@@ -43,5 +43,7 @@ return [
     'mark_confirm' => 'Confirm done',
     'mark_text' => 'Are you sure you have done all the the work? Trust is needed there.',
 
+    'comment' => 'Comment',
+
 
 ];