+++ resources/lang/es/challenges.php
@@ -43,4 +43,6 @@ return [
     'mark_confirm' => 'Confirmo la superación',
     'mark_text' => '¿Estás seguro/a que has superado la tarea? La confianza es necesaria.',
 
+    'comment' => 'Comenta',
+
 ];