+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## v1.10 Beta (XX/10/2020)
+## v1.10 Beta (21/10/2020)
 ### Added
 - Import from Google Classroom.
 - Cookie consent improved.